+++ src/components/Table/Table.tsx
@@ -1,9 +1,22 @@
 import 'react-dates/initialize';
 import React, { ReactNode, SyntheticEvent, Component } from 'react';
 import { CSSProperties } from 'styled-components';
+import { IoIosClose } from 'react-icons/io';
+import { Field, Formik, FormikProps } from 'formik';
+import { ObjectSchema } from 'yup';
+import ErrorField from '../Input/ErrorField';
+import FormikInput from '../Input/FormikInput';
+import colors from '../../theme/colors';
 import Grid from '../Grid';
 import Loader from '../Loader';
-import { AnalyticsProps } from '../../analytics';
+import {
+  AnalyticsComponent,
+  AnalyticsProps,
+  AnalyticsEvents,
+  AnalyticsComponentType,
+} from '../../analytics';
+
+// styled components
 import {
   ButtonTableRow,
   LoaderContainer,
@@ -15,6 +28,10 @@ import {
   TableHeaderCell,
   TableWithHeaderSectionContainer,
   TableWrapper,
+  StyledReactIcon,
+  StyledIcon,
+  StyledForm,
+  FormWrapper,
 } from './styled';
 
 interface Data {
@@ -27,13 +44,35 @@ type Render = {
   rowValues: Data;
 };
 
-type Column = {
+export interface ColumnFunctionProps {
+  value: string;
+  currentValue: string;
+  id: string;
+}
+
+export interface EditableFields {
+  fn: (props: ColumnFunctionProps) => void;
+  validationSchema: ObjectSchema<any>;
+}
+
+export type Column = {
   title: string;
   key: string;
-  render: ({ value, rowValues }: Render) => ReactNode;
+  render?: ({ value, rowValues }: Render) => ReactNode;
   style?: CSSProperties;
   alignRight?: any;
+  editable?: EditableFields;
 };
+
+interface FormValeus {
+  rowField: string;
+}
+interface EditIconsProps {
+  field: string;
+  rowId: string;
+  fn?: ({ value, currentValue, id }: ColumnFunctionProps) => any;
+  validationSchema?: ObjectSchema<any>;
+}
 
 type RowActions<T> =
   | Array<{
@@ -62,7 +101,7 @@ interface TableButtonProps extends AnalyticsProps {
 
 type Props<T = Data> = {
   columns: Array<Column> | any;
-  data: Array<any>;
+  data: Array<Data>;
   className?: string;
   style?: CSSProperties;
   options: TableOptions<T>;
@@ -70,6 +109,7 @@ type Props<T = Data> = {
   renderEmptyTable?: boolean;
   tableButton?: Array<TableButtonProps>;
   actionsRowTitle?: string;
+  editableId?: string;
 } & AnalyticsProps;
 
 type State = {
@@ -77,6 +117,7 @@ type State = {
     [key in number | string]: any;
   };
   selectedAll: boolean;
+  currentlyModifyingRowId?: string | null;
 };
 
 const rowActionsExist = (actions: RowActions<any>) =>
@@ -104,6 +145,7 @@ class Table<T> extends Component<Props<T>, State> {
   state: State = {
     selected: {},
     selectedAll: false,
+    currentlyModifyingRowId: null,
   };
 
   toggleRow = (id: number | string) => () => {
@@ -209,25 +251,145 @@ class Table<T> extends Component<Props<T>, State> {
       </TableHeaderCell>
     ));
 
+  renderEditIcons = ({ field, rowId, fn, validationSchema }: EditIconsProps) => {
+    const { currentlyModifyingRowId } = this.state;
+    if (currentlyModifyingRowId === rowId) {
+      return (
+        <Formik
+          initialValues={{ rowField: field }}
+          onSubmit={({ rowField }) => {
+            this.setState({
+              currentlyModifyingRowId: null,
+            });
+            fn?.({ value: field, currentValue: rowField, id: rowId });
+          }}
+          validationSchema={validationSchema}
+        >
+          {({ isValid, values: { rowField } }: FormikProps<FormValeus>) => (
+            <StyledForm>
+              <AnalyticsComponent>
+                {({ pages, tracker, merchantId, ip }) => (
+                  <FormWrapper>
+                    <Field
+                      tag="editable_text_input"
+                      type="text"
+                      name="rowField"
+                      component={FormikInput}
+                    />
+                    <StyledReactIcon
+                      data-testid="save"
+                      color={colors.green}
+                      onClick={() => {
+                        tracker.track({
+                          event: AnalyticsEvents.CLICK,
+                          type: AnalyticsComponentType.ICON,
+                          tag: 'icon_save',
+                          pages,
+                          merchantId,
+                          ip,
+                        });
+                        if (isValid) {
+                          fn?.({ value: field, currentValue: rowField, id: rowId });
+                          this.setState({
+                            currentlyModifyingRowId: null,
+                          });
+                        }
+                      }}
+                    />
+                    <StyledReactIcon
+                      data-testid="cancel"
+                      as={IoIosClose}
+                      color={colors.red}
+                      onClick={() => {
+                        tracker.track({
+                          event: AnalyticsEvents.CLICK,
+                          type: AnalyticsComponentType.ICON,
+                          tag: 'icon_cancel',
+                          pages,
+                          merchantId,
+                          ip,
+                        });
+                        this.setState({
+                          currentlyModifyingRowId: null,
+                        });
+                      }}
+                    />
+                  </FormWrapper>
+                )}
+              </AnalyticsComponent>
+              <ErrorField name="rowField" />
+            </StyledForm>
+          )}
+        </Formik>
+      );
+    }
+    return (
+      <AnalyticsComponent>
+        {({ pages, tracker, merchantId, ip }) => (
+          <>
+            {field}
+            <StyledIcon
+              onClick={() => {
+                tracker.track({
+                  event: AnalyticsEvents.CLICK,
+                  type: AnalyticsComponentType.ICON,
+                  tag: 'icon_edit',
+                  pages,
+                  merchantId,
+                  ip,
+                });
+                this.setState({
+                  currentlyModifyingRowId: rowId,
+                });
+              }}
+            />
+          </>
+        )}
+      </AnalyticsComponent>
+    );
+  };
+
   renderData = (columns: Array<Column>, data: Array<Data>) => {
     const newColumns = this.generateColumns(columns);
     const newData = this.generateRows(data);
 
-    const { tableButton } = this.props;
+    const { tableButton, editableId = '0' } = this.props;
 
     return newData.map((row) => (
       <TableRow key={row.id} noBottomBorder={!tableButton}>
-        {newColumns.map((column: Column, index: number) => (
-          <TableCell
-            key={index}
-            isActionsCell={column.key === 'actions'}
-            style={column.style || {}}
-          >
-            {column.render
-              ? column.render({ value: row[column.key], rowValues: row })
-              : row[column.key]}
-          </TableCell>
-        ))}
+        {newColumns.map(({ render, key, editable, style }: Column, index: number) => {
+          const isEditable = editable
+            ? this.renderEditIcons({
+                field: row[key],
+                rowId: row[editableId],
+                fn: editable.fn,
+                validationSchema: editable.validationSchema,
+              })
+            : row[key];
+
+          const hasRenderFn = render ? render({ value: row[key], rowValues: row }) : isEditable;
+
+          const hasRenderAndEditable =
+            render && editable
+              ? this.renderEditIcons({
+                  field: render({ value: row[key], rowValues: row }) as string,
+                  rowId: row[editableId],
+                  fn: editable.fn,
+                  validationSchema: editable.validationSchema,
+                })
+              : hasRenderFn;
+
+          return (
+            <TableCell
+              isEditableCell={!!editable}
+              key={index}
+              isActionsCell={key === 'actions'}
+              style={style || {}}
+            >
+              {hasRenderAndEditable}
+            </TableCell>
+          );
+        })}
       </TableRow>
     ));
   };