+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-# [2.5.21] - 24.01.2022
+# [2.5.22] - 04.02.2022
+
+## Added
+
+- Option in column table to make fields(columns) editable
+- onBlur function to trigger persa when input and textarea are focused-out
+
+# [2.5.21] - 25.01.2022
 
 ## Change